backend/google-auth/requirements.txt
pydantic==2.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0